refractored cryptography functions into one module
